+++ lib/src/bitflag_base.dart
@@ -2,31 +2,31 @@ part of '../bit_flag.dart';
 
 @protected
 abstract interface class IBitFlag {
-  /// Returns the value of the flag, if the class is [FlipBitFlag], the value is always zero.
+  /// Returns the value of the flag, if the class is [ZeroBitFlag], the value is always zero.
   int get value;
 
   /// OR (Bitwise) two values of this and [other].
-  /// If both this and [other] are [FlipBitFlag], the result will be a [FlipBitFlag] too.
+  /// If both this and [other] are [ZeroBitFlag], the result will be a [ZeroBitFlag] too.
   IBitFlag operator |(covariant IBitFlag other);
 
   /// AND (Bitwise) two values of this and [other].
-  /// If both this and [other] are [FlipBitFlag], the result will be a [FlipBitFlag] too.
+  /// If both this and [other] are [ZeroBitFlag], the result will be a [ZeroBitFlag] too.
   IBitFlag operator &(covariant IBitFlag other);
 
   /// XOR (Bitwise) two values of this and [other].
-  /// If both this and [other] are [FlipBitFlag], the result will be a [FlipBitFlag] too.
+  /// If both this and [other] are [ZeroBitFlag], the result will be a [ZeroBitFlag] too.
   IBitFlag operator ^(covariant IBitFlag other);
 
   /// SHIFT LEFT (Bitwise) two values of this and [other].
-  /// If both this and [other] are [FlipBitFlag], the result will be a [FlipBitFlag] too.
+  /// If both this and [other] are [ZeroBitFlag], the result will be a [ZeroBitFlag] too.
   IBitFlag operator <<(int other);
 
   /// SHIFT RIGHT (Bitwise) two values of this and [other].
-  /// If both this and [other] are [FlipBitFlag], the result will be a [FlipBitFlag] too.
+  /// If both this and [other] are [ZeroBitFlag], the result will be a [ZeroBitFlag] too.
   IBitFlag operator >>(int other);
 
   /// Returns true if the [origin] contains the [flag].
-  /// If the [flag] is a [FlipBitFlag], the result will be true when the [origin] does not contain [flag.otherSide].
+  /// If the [flag] is a [ZeroBitFlag], the result will be true when the [origin] does not contain [flag.mask].
   bool hasFlag(IBitFlag flag);
 
   /// Returns the value of [origin] as String with the [radix].
@@ -36,12 +36,12 @@ abstract interface class IBitFlag {
   /// If the [width] is not set, the bit length will be auto-set.
   String toBinaryString([int width]);
 
-  /// Flip the flag. If the [origin] is a [FlipBitFlag], the result will be a normal [BitFlag].
-  /// Otherwise, the result will be changed to [FlipBitFlag].
-  IBitFlag toFlip();
+  /// Flip the flag. If the [origin] is a [ZeroBitFlag], the result will be a normal [BitFlag].
+  /// Otherwise, the result will be changed to [ZeroBitFlag].
+  IBitFlag flip();
 
   /// Returns flag and keeps the bit that matches the 1-bit in the [mask]. All of the other bits will be set to 0.
-  IBitFlag extract(IBitFlag mask);
+  IBitFlag extract(IBitFlag mask, {bool trimRight});
 
   /// Toggle the bit that matches the 1-bit in the mask.
   IBitFlag toggle(IBitFlag mask);
@@ -49,16 +49,16 @@ abstract interface class IBitFlag {
 
 @immutable
 // ignore: missing_override_of_must_be_overridden
-abstract class FlipBitFlag implements IBitFlag {
-  const FlipBitFlag();
+abstract class ZeroBitFlag implements IBitFlag {
+  const ZeroBitFlag();
 
-  IBitFlag get otherSide;
+  IBitFlag get mask;
 
   @override
   int get value => 0;
 
   @override
   String toString() {
-    return 'BitFlag($value┆${otherSide.value})';
+    return 'BitFlag($value┆${mask.value})';
   }
 }